вынес projects в отдельную папку, куда отдельным файлом занес функцию, нормализующую стейт. Импортировал их в компонент App
