+++ src/components/App/App.js
@@ -8,84 +8,20 @@ import ThemeSwitcher from '../UI/ThemeSwitcher/ThemeSwitcher'
 // Импорт контекста
 import { DEFAULT_THEME, ThemeContext } from "./ThemeContext"
 
+import projects from '../ProjectsData/projectsData'
+import normalizeState from '../ProjectsData/stateNormalizer'
+
 import classes from './App.scss'
 import classnames from "classnames/bind"
 const cx = classnames.bind(classes)
-// import { createGlobalStyle } from 'styled-components'
+
+const {projectsById, tasksById} = normalizeState(projects)
 
 class App extends React.Component {
   state = {
     theme: DEFAULT_THEME,
-    projects: [
-      {
-        id: 1,
-        name: 'Main stuff',
-        description: 'Here are a few tasks that should be completed this month',
-        tasks: [
-          {
-            id: 1,
-            name: 'Buy milk',
-            description: '20 packages of Parmalat 1.5%',
-            completed: true
-          },
-          {
-            id: 2,
-            name: 'Write a review',
-            description: 'Write a review for season 1 of the series "Love Death + Robots"',
-            completed: false
-          },
-          {
-            id: 3,
-            name: 'Find ball and sneakers',
-            description: "We'll play basketball this Wednesday",
-            completed: true
-          },
-          {
-            id: 4,
-            name: 'Take a pizza cooking lesson',
-            description: 'Ur gf bet she cooks better🤢',
-            completed: true
-          },
-          {
-            id: 5,
-            name: 'Visit parents',
-            description: 'This weekend we might go for a walk. Go visit a restaurant on Kitay-Gorod',
-            completed: true
-          }
-        ]
-      },
-      { 
-        id: 2,
-        name: 'University',
-        description: 'Here are the university tasks that need to be done',
-        tasks: [
-          {
-            id: 6,
-            name: 'IT-infrastructure',
-            description: '5th Lab. 23:59 LMS May 26',
-            completed: true
-          },
-          {
-            id: 7,
-            name: 'Strategic management',
-            description: 'The 2nd group project was listed. We need to show it until 30th of May',
-            completed: false
-          },
-          {
-            id: 8,
-            name: 'Find mates for strategic management',
-            description: "2nd group project is coming soon",
-            completed: true
-          },
-          {
-            id: 9,
-            name: 'ImProfEdu',
-            description: '5th HW – take a video',
-            completed: true
-          }
-        ]
-      }
-    ]
+    projectsById,
+    tasksById
   };
 
   // Смена статуса таски completed: done / undone
@@ -126,48 +62,7 @@ class App extends React.Component {
   }
 
   render() {
-    // Нормализация стейта
-    const normalizeState = projectsArray => { 
-      const normalizedProjectsArray = {} // проекты с айдишниками тасок
-      const normalizedTasksArray = {} // все таски 
-      
-      const normilizedState = { 
-        projectsById: normalizedProjectsArray,
-        tasksById: normalizedTasksArray
-      }
-
-      // Получить ID тасок по проекту
-      const getTasksIdsByProject = tasks => {
-        const projectTasksIds = []
-        tasks.map(task => projectTasksIds.push(task.id))
-        return projectTasksIds
-      }
-
-      projectsArray.map( project => {
-        const projectTasks = project.tasks
-        projectTasks.map( task => {
-          return normalizedTasksArray[task.id] = {
-            id: task.id,
-            name: task.name,
-            description: task.description,
-            completed: task.completed
-          }
-        })
-        return normalizedProjectsArray[project.id] = {
-          id: project.id,
-          name: project.name,
-          tasksIds: getTasksIdsByProject(project.tasks)
-        }
-      })
-
-      return normilizedState
-    };
-
-
-    const {projectsById: projects, tasksById: tasks} = normalizeState(this.state.projects)
-    console.log('PROJECTS', projects);
-    console.log('TASKS', tasks);
-
+    console.log(this.state);
     return (
       <section className={cx('application-wrapper', `application-wrapper-theme-${this.state.theme}`)}>
         <div className={cx('tasks-wrapper__layout')}>
