+++ src/components/ProjectsData/projectsData.js
@@ -1,0 +1,72 @@
+const projects = [
+    {
+      id: 1,
+      name: 'Main stuff',
+      description: 'Here are a few tasks that should be completed this month',
+      tasks: [
+        {
+          id: 1,
+          name: 'Buy milk',
+          description: '20 packages of Parmalat 1.5%',
+          completed: true
+        },
+        {
+          id: 2,
+          name: 'Write a review',
+          description: 'Write a review for season 1 of the series "Love Death + Robots"',
+          completed: false
+        },
+        {
+          id: 3,
+          name: 'Find ball and sneakers',
+          description: "We'll play basketball this Wednesday",
+          completed: true
+        },
+        {
+          id: 4,
+          name: 'Take a pizza cooking lesson',
+          description: 'Ur gf bet she cooks better🤢',
+          completed: true
+        },
+        {
+          id: 5,
+          name: 'Visit parents',
+          description: 'This weekend we might go for a walk. Go visit a restaurant on Kitay-Gorod',
+          completed: true
+        }
+      ]
+    },
+    { 
+      id: 2,
+      name: 'University',
+      description: 'Here are the university tasks that need to be done',
+      tasks: [
+        {
+          id: 6,
+          name: 'IT-infrastructure',
+          description: '5th Lab. 23:59 LMS May 26',
+          completed: true
+        },
+        {
+          id: 7,
+          name: 'Strategic management',
+          description: 'The 2nd group project was listed. We need to show it until 30th of May',
+          completed: false
+        },
+        {
+          id: 8,
+          name: 'Find mates for strategic management',
+          description: "2nd group project is coming soon",
+          completed: true
+        },
+        {
+          id: 9,
+          name: 'ImProfEdu',
+          description: '5th HW – take a video',
+          completed: true
+        }
+      ]
+    }
+  ]
+
+  export default projects
